Add command handling
